requirements.txt
gradio~=3.43.2
transformers>=4.31.0
similarities>=1.1.7
sentencepiece
markdown
PyPDF2
python-docx
pandas
protobuf
loguru
peft